Move ControllableDeployer to flocker.node.testtools
