+++ flocker/node/testtools.py
@@ -10,7 +10,10 @@ import pwd
 import socket
 from unittest import skipUnless
 
+from zope.interface import implementer
+
 from ._docker import BASE_DOCKER_API_URL
+from . import IDeployer
 from ..testtools import loop_until
 
 DOCKER_SOCKET_PATH = BASE_DOCKER_API_URL.split(':/')[-1]
@@ -67,3 +70,24 @@ def wait_for_unit_state(docker_client, unit_name, expected_activation_states):
         return responded
 
     return loop_until(check_if_in_states)
+
+
+@implementer(IDeployer)
+class ControllableDeployer(object):
+    """
+    ``IDeployer`` whose results can be controlled.
+    """
+    def __init__(self, local_states, calculated_actions):
+        self.local_states = local_states
+        self.calculated_actions = calculated_actions
+        self.calculate_inputs = []
+
+    def discover_local_state(self):
+        return self.local_states.pop(0)
+
+    def calculate_necessary_state_changes(self, local_state,
+                                          desired_configuration,
+                                          cluster_state):
+        self.calculate_inputs.append(
+            (local_state, desired_configuration, cluster_state))
+        return self.calculated_actions.pop(0)
